+++ Src/Knowledge/AllSpellsShaman.lua
@@ -11,8 +11,8 @@ local spellIdsModule = BomModuleManager.spellIdsModule
 ---@param allBuffs BomBuffDefinition[]
 ---@param enchants BomEnchantmentsMapping
 function shamanModule:SetupShamanSpells(allBuffs, enchants)
-  local duration = allBuffsModule.TbcOrClassic(allBuffsModule.DURATION_20M, allBuffsModule.DURATION_10M)
-  local enchantmentDuration = allBuffsModule.TbcOrClassic(allBuffsModule.DURATION_30M, allBuffsModule.DURATION_5M) -- TBC: Shaman enchants become 30min
+  local duration = allBuffsModule.TbcOrClassic(allBuffsModule.TWENTY_MINUTES, allBuffsModule.TEN_MINUTES)
+  local enchantmentDuration = allBuffsModule.TbcOrClassic(allBuffsModule.HALF_AN_HOUR, allBuffsModule.FIVE_MINUTES) -- TBC: Shaman enchants become 30min
 
   -- Flametongue Weapon
   buffDefModule:createAndRegisterBuff(allBuffs, spellIdsModule.Shaman_Flametongue6, nil)
@@ -114,7 +114,7 @@ function shamanModule:SetupShamanSpells(allBuffs, enchants)
   -- Water Walking (no reagent in WotLK with a glyph)
   buffDefModule:createAndRegisterBuff(allBuffs, 546, nil)
                :IsDefault(false)
-               :SingleDuration(allBuffsModule.DURATION_10M)
+               :SingleDuration(allBuffsModule.TEN_MINUTES)
                :DefaultTargetClasses(allBuffsModule.BOM_ALL_CLASSES)
                :RequirePlayerClass("SHAMAN")
                :Category("class")
@@ -122,7 +122,7 @@ function shamanModule:SetupShamanSpells(allBuffs, enchants)
   -- Water Breathing (no reagent in WotLK with a glyph)
   buffDefModule:createAndRegisterBuff(allBuffs, 131, nil)
                :IsDefault(false)
-               :SingleDuration(allBuffsModule.DURATION_10M)
+               :SingleDuration(allBuffsModule.TEN_MINUTES)
                :DefaultTargetClasses(allBuffsModule.BOM_ALL_CLASSES)
                :RequirePlayerClass("SHAMAN")
                :Category("class")